refactor : BD-31 WorkSpaceCell로 리스트 분류
